+++ modules/container-web-app/r-appservice.tf
@@ -115,8 +115,12 @@ resource "azurerm_linux_web_app" "app_service_linux_container" {
   client_certificate_enabled = var.client_certificate_enabled
   https_only                 = var.https_only
 
-  identity {
-    type = "SystemAssigned"
+  dynamic "identity" {
+    for_each = var.identity[*]
+    content {
+      type         = var.identity.type
+      identity_ids = var.identity.identity_ids
+    }
   }
 
   dynamic "backup" {
@@ -305,8 +309,12 @@ resource "azurerm_linux_web_app_slot" "app_service_linux_container_slot" {
   client_affinity_enabled = var.client_affinity_enabled
   https_only              = var.https_only
 
-  identity {
-    type = "SystemAssigned"
+  dynamic "identity" {
+    for_each = var.identity[*]
+    content {
+      type         = var.identity.type
+      identity_ids = var.identity.identity_ids
+    }
   }
 
   dynamic "storage_account" {