+++ modules/windows-web-app/r-appservice.tf
@@ -126,8 +126,12 @@ resource "azurerm_windows_web_app" "app_service_windows" {
   client_certificate_enabled = var.client_certificate_enabled
   https_only                 = var.https_only
 
-  identity {
-    type = "SystemAssigned"
+  dynamic "identity" {
+    for_each = var.identity[*]
+    content {
+      type         = var.identity.type
+      identity_ids = var.identity.identity_ids
+    }
   }
 
   dynamic "backup" {