+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+    * [GH-3](https://github.com/claranet/terraform-azurerm-app-service/pull/3): Add Support to User Assigned Identities
+
 # v7.4.0 - 2023-08-11
 
 Added